Responsive v1

+++ src/Entity/Burger.php
@@ -21,7 +21,7 @@ class Burger
     private ?string $description = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $price = null;
+    private ?float $price = null;
 
     #[ORM\Column(length: 255, nullable: true)]
     private ?string $img = null;
@@ -70,12 +70,12 @@ class Burger
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(?int $price): self
+    public function setPrice(?float $price): self
     {
         $this->price = $price;
